Renames RecipeParserDeps -> RecipeReaderDeps
